Remove circular reference
