+++ tests/integrationv2/conftest.py
@@ -3,13 +3,11 @@ from global_flags import set_flag, S2N_PROVIDER_VERSION, S2N_FIPS_MODE, S2N_NO_P
 
 
 def pytest_addoption(parser):
-    parser.addoption("--provider-version", action="store", dest="provider-version",
-                     default=None, type=str, help="Set the version of the TLS provider")
-    parser.addoption("--fips-mode", action="store", dest="fips-mode",
-                     default=False, type=int, help="S2N is running in FIPS mode")
-    parser.addoption("--no-pq", action="store", dest="no-pq",
-                     default=False, type=int, help="Turn off PQ support")
-
+    parser.addoption("--provider-version", action="store", dest="provider-version", default=None, type=str, help="Set the version of the TLS provider")
+    parser.addoption("--fips-mode", action="store", dest="fips-mode", default=False, type=int, help="S2N is running in FIPS mode")
+    parser.addoption("--no-pq", action="store", dest="no-pq", default=False, type=int, help="Turn off PQ support")
+    parser.addoption("--provider-criterion", action="store", dest="provider-criterion",
+      default=False, type=int, help="0=off(default, 1=Run the test under a Criterion provider in delta mode, 2=Run in baseline mode.")
 
 def pytest_configure(config):
     """
@@ -22,12 +20,14 @@ def pytest_configure(config):
 
     no_pq = config.getoption('no-pq', 0)
     fips_mode = config.getoption('fips-mode', 0)
-    if no_pq == 1:
+    use_criterion = config.getoption('proivder-criterion', 0)
+    if no_pq is 1:
         set_flag(S2N_NO_PQ, True)
-    if fips_mode == 1:
+    if fips_mode is 1:
         set_flag(S2N_FIPS_MODE, True)
 
     set_flag(S2N_PROVIDER_VERSION, config.getoption('provider-version', None))
+    set_flag(S2N_USECRITERION, config.getoption('provider-criterion', None))
 
 
 def pytest_collection_modifyitems(config, items):
