+++ tests/integrationv2/providers.py
@@ -13,6 +13,22 @@ TLS_13_LIBCRYPTOS = {
     "openssl-1.1.1"
 }
 
+def get_expected_s2n_version(protocol, provider):
+    """
+    s2nd and s2nc print a number for the negotiated TLS version.
+
+    provider is s2n's peer. If s2n tries to speak to s2n < tls13,
+    tls12 is always chosen. This is true even when the requested
+    protocol is less than tls12.
+    """
+    if (issubclass(provider, S2N)) and protocol != Protocols.TLS13:
+        version = '33'
+    else:
+        version = protocol.value
+
+    return version
+
+
 
 class Provider(object):
     """
