+++ tests/integrationv2/test_client_authentication.py
@@ -7,8 +7,8 @@ from configuration import (available_ports, ALL_TEST_CIPHERS, ALL_TEST_CURVES,
                            ALL_TEST_CERTS, PROTOCOLS)
 from common import Certificates, ProviderOptions, Protocols, data_bytes
 from fixtures import managed_process
-from providers import Provider, S2N, OpenSSL
-from utils import invalid_test_parameters, get_parameter_name, get_expected_s2n_version, to_bytes
+from providers import Provider, S2N, OpenSSL, get_expected_s2n_version
+from utils import invalid_test_parameters, get_parameter_name, to_bytes
 
 
 # If we test every available cert, the test takes too long.
