+++ tests/integrationv2/test_dynamic_record_sizes.py
@@ -7,8 +7,8 @@ import time
 from configuration import available_ports, ALL_TEST_CIPHERS, ALL_TEST_CURVES, ALL_TEST_CERTS, PROVIDERS, PROTOCOLS
 from common import ProviderOptions, data_bytes, Protocols
 from fixtures import managed_process, custom_mtu
-from providers import Provider, S2N, OpenSSL, Tcpdump
-from utils import invalid_test_parameters, get_parameter_name, get_expected_s2n_version, to_bytes
+from providers import Provider, S2N, OpenSSL, Tcpdump, get_expected_s2n_version
+from utils import invalid_test_parameters, get_parameter_name, to_bytes
 
 
 def find_fragmented_packet(results):
