+++ tests/integrationv2/test_version_negotiation.py
@@ -4,8 +4,8 @@ import pytest
 from configuration import available_ports, ALL_TEST_CIPHERS, ALL_TEST_CURVES, ALL_TEST_CERTS
 from common import ProviderOptions, Protocols, data_bytes
 from fixtures import managed_process
-from providers import Provider, S2N, OpenSSL, GnuTLS
-from utils import invalid_test_parameters, get_parameter_name, get_expected_s2n_version, get_expected_openssl_version, \
+from providers import Provider, S2N, OpenSSL, GnuTLS, get_expected_s2n_version
+from utils import invalid_test_parameters, get_parameter_name, get_expected_openssl_version, \
     to_bytes, get_expected_gnutls_version
