+++ tests/integrationv2/utils.py
@@ -1,29 +1,11 @@
 import os
 from common import Protocols, Curves, Ciphers
-from providers import S2N, OpenSSL
 from global_flags import get_flag, S2N_FIPS_MODE, S2N_PROVIDER_VERSION
 
 EXECUTABLE=33261
 
 def to_bytes(val):
     return bytes(str(val).encode('utf-8'))
-
-
-def get_expected_s2n_version(protocol, provider):
-    """
-    s2nd and s2nc print a number for the negotiated TLS version.
-
-    provider is s2n's peer. If s2n tries to speak to s2n < tls13,
-    tls12 is always chosen. This is true even when the requested
-    protocol is less than tls12.
-    """
-    if (issubclass(provider, S2N)) and protocol != Protocols.TLS13:
-        version = '33'
-    else:
-        version = protocol.value
-
-    return version
-
 
 def get_expected_openssl_version(protocol):
     return {
